fix: carry replica ids as u16 in benches, examples, and loopback tests
The u16 ReplicaId widening reached the library and its unit tests but not the
bench, example, and driver loopback targets, which still constructed replica
ids from u8 loop and helper variables — so `cargo build --all-targets` (and
CI's all-target build) failed even though the library tests passed. Construct
each replica id as a u16 across those targets.

+++ viewstamp-compio/tests/batching_loopback.rs
@@ -124,14 +124,14 @@ impl StateMachine for SharedSm {
 
 fn mk_dialer(me: u8) -> Rc<dyn Fn(Peer) -> Conn<Labeled<Passthrough>>> {
   Rc::new(move |_peer| {
-    let opts = LabelOptions::new(CLUSTER, Peer::Replica(ReplicaId::new(me)));
+    let opts = LabelOptions::new(CLUSTER, Peer::Replica(ReplicaId::new(me as u16)));
     Conn::from_parts(Labeled::dialer(Passthrough::new(), &opts))
   })
 }
 
 fn mk_acceptor(me: u8) -> Rc<dyn Fn() -> Conn<Labeled<Passthrough>>> {
   Rc::new(move || {
-    let opts = LabelOptions::new(CLUSTER, Peer::Replica(ReplicaId::new(me)));
+    let opts = LabelOptions::new(CLUSTER, Peer::Replica(ReplicaId::new(me as u16)));
     Conn::from_parts(Labeled::acceptor(Passthrough::new(), &opts))
   })
 }
@@ -152,9 +152,9 @@ async fn spawn_cluster(
   for id in 0u8..3 {
     let peers: Vec<_> = (0u8..3)
       .filter(|&p| p != id)
-      .map(|p| (ReplicaId::new(p), addrs[p as usize]))
+      .map(|p| (ReplicaId::new(p as u16), addrs[p as usize]))
       .collect();
-    let config = viewstamp_proto::Config::try_new(CLUSTER, ReplicaId::new(id), 3).unwrap();
+    let config = viewstamp_proto::Config::try_new(CLUSTER, ReplicaId::new(id as u16), 3).unwrap();
     let (ready_tx, ready_rx) = flume::unbounded();
     let wal = Notifying::new(InMemoryWal::new(), ready_tx.clone());
     let sb = Notifying::new(InMemorySuperblock::new(), ready_tx);
